+++ care/emr/tests/test_reset_password_api.py
@@ -457,48 +457,3 @@ class ResetPasswordAPITest(CareAPITestBase):
                 format="json",
             )
             self.assertEqual(response.status_code, 400)
-
-    def test_change_password_with_leading_whitespace(self):
-        """
-        Test that password with leading whitespace is handled consistently.
-        The password should be stripped before validation, matching login behavior.
-        """
-        self.client.force_authenticate(user=self.user)
-        new_password = "newpassword@123"
-        response = self.client.put(
-            self.change_password_url,
-            {"old_password": f"  {self.password}", "new_password": new_password},
-            format="json",
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.data, {"message": "Password updated successfully"})
-
-    def test_change_password_with_trailing_whitespace(self):
-        """
-        Test that password with trailing whitespace is handled consistently.
-        The password should be stripped before validation, matching login behavior.
-        """
-        self.client.force_authenticate(user=self.user)
-        new_password = "newpassword@123"
-        response = self.client.put(
-            self.change_password_url,
-            {"old_password": f"{self.password}  ", "new_password": new_password},
-            format="json",
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.data, {"message": "Password updated successfully"})
-
-    def test_change_password_with_leading_and_trailing_whitespace(self):
-        """
-        Test that password with both leading and trailing whitespace is handled consistently.
-        The password should be stripped before validation, matching login behavior.
-        """
-        self.client.force_authenticate(user=self.user)
-        new_password = "newpassword@123"
-        response = self.client.put(
-            self.change_password_url,
-            {"old_password": f"  {self.password}  ", "new_password": new_password},
-            format="json",
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.data, {"message": "Password updated successfully"})
